Move icons and not-found page to shared-ui library
