+++ libs/shared-ui/src/lib/icons/chevron-up/chevron-up-icon.component.ts
