+++ libs/shared-ui/src/lib/icons/coins/coins-icon.component.ts
