+++ libs/shared-ui/src/lib/icons/floppy-disk/floppy-disk-icon.component.ts
