+++ libs/shared-ui/src/lib/icons/icon.scss
