+++ libs/shared-ui/src/lib/icons/stopwatch/stopwatch-icon.component.ts
