+++ libs/shared-ui/src/lib/icons/trash/trash-icon.component.ts
