+++ libs/shared-ui/src/lib/icons/weight-hanging/weight-hanging-icon.component.ts
