+++ libs/shared-ui/src/lib/not-found/not-found.page.scss
